+++ apps/ledger-live-desktop/src/internal/commands/index.js
@@ -24,7 +24,9 @@ import websocketBridge from "./websocketBridge";
 import checkSignatureAndPrepare from "./checkSignatureAndPrepare";
 import getTransactionId from "./getTransactionId";
 import scanDescriptors from "./scanDescriptors";
+import installLanguage from "./installLanguage";
 import getAppAndVersion from "./getAppAndVersion";
+import getDeviceInfo from "./getDeviceInfo";
 import { commands as bridgeProxyCommands } from "~/renderer/bridge/proxy-commands";
 
 export const commandsById = {
@@ -53,8 +55,10 @@ export const commandsById = {
   testInterval,
   websocketBridge,
   scanDescriptors,
+  installLanguage,
   signMessage,
   getAppAndVersion,
+  getDeviceInfo,
 };
 
 export type Commands = typeof commandsById;