+++ apps/ledger-live-mobile/src/components/RootNavigator/BaseNavigator.tsx
@@ -9,6 +9,7 @@ import { useTranslation } from "react-i18next";
 import { Flex, Icons } from "@ledgerhq/native-ui";
 import { useTheme } from "styled-components/native";
 import useFeature from "@ledgerhq/live-common/featureFlags/useFeature";
+import { useSelector } from "react-redux";
 import { ScreenName, NavigatorName } from "../../const";
 import * as families from "../../families";
 import OperationDetails, {
@@ -58,6 +59,8 @@ import AnalyticsOperations from "../../screens/Analytics/Operations";
 import NftNavigator from "./NftNavigator";
 import { getStackNavigatorConfig } from "../../navigation/navigatorConfig";
 import Account from "../../screens/Account";
+// eslint-disable-next-line import/no-cycle
+import ReadOnlyAccount from "../../screens/Account/ReadOnly/ReadOnlyAccount";
 import TransparentHeaderNavigationOptions from "../../navigation/TransparentHeaderNavigationOptions";
 import styles from "../../navigation/styles";
 import HeaderRightClose from "../HeaderRightClose";
@@ -70,11 +73,6 @@ import PlatformApp from "../../screens/Platform/App";
 import AccountsNavigator from "./AccountsNavigator";
 
 import MarketCurrencySelect from "../../screens/Market/MarketCurrencySelect";
-import SwapFormSelectAccount from "../../screens/Swap/FormSelection/SelectAccountScreen";
-import SwapFormSelectCurrency from "../../screens/Swap/FormSelection/SelectCurrencyScreen";
-import SwapFormSelectFees from "../../screens/Swap/FormSelection/SelectFeesScreen";
-import SwapFormSelectProviderRate from "../../screens/Swap/FormSelection/SelectProviderRateScreen";
-import SwapOperationDetails from "../../screens/Swap/OperationDetails";
 
 import ProviderList from "../../screens/Exchange/ProviderList";
 import ProviderView from "../../screens/Exchange/ProviderView";
@@ -97,8 +95,9 @@ import {
   BleDevicePairingFlow,
   BleDevicePairingFlowParams,
 } from "../../screens/BleDevicePairingFlow/index";
+import { readOnlyModeEnabledSelector } from "../../reducers/settings";
+import { accountsSelector } from "../../reducers/accounts";
 
-// TODO: types for each screens and navigators need to be set
 export type BaseNavigatorStackParamList = {
   BleDevicePairingFlow: BleDevicePairingFlowParams;
 
@@ -123,12 +122,15 @@ export default function BaseNavigator() {
   const ptxSmartRoutingMobile = useFeature("ptxSmartRoutingMobile");
   const walletConnectLiveApp = useFeature("walletConnectLiveApp");
   const noNanoBuyNanoWallScreenOptions = useNoNanoBuyNanoWallScreenOptions();
+  const accounts = useSelector(accountsSelector);
+  const readOnlyModeEnabled =
+    useSelector(readOnlyModeEnabledSelector) && accounts.length <= 0;
 
   return (
     <Stack.Navigator
       screenOptions={{
         ...stackNavigationConfig,
-        ...TransitionPresets.ModalPresentation,
+        ...TransitionPresets.DefaultTransition,
       }}
     >
       <Stack.Screen
@@ -239,63 +241,7 @@ export default function BaseNavigator() {
       <Stack.Screen
         name={NavigatorName.Swap}
         component={SwapNavigator}
-        options={{
-          ...stackNavigationConfig,
-          headerLeft: () => null,
-          title: t("transfer.swap.form.tab"),
-        }}
-      />
-      <Stack.Screen
-        name={ScreenName.SwapV2FormSelectAccount}
-        component={SwapFormSelectAccount}
-        options={({ route }) => ({
-          headerTitle: () => (
-            <StepHeader
-              title={
-                route.params.target === "from"
-                  ? t("transfer.swap.form.from")
-                  : t("transfer.swap.form.to")
-              }
-            />
-          ),
-          headerRight: () => null,
-        })}
-      />
-      <Stack.Screen
-        name={ScreenName.SwapOperationDetails}
-        component={SwapOperationDetails}
-        options={{
-          title: t("transfer.swap.form.tab"),
-          headerRight: () => null,
-        }}
-      />
-      <Stack.Screen
-        name={ScreenName.SwapV2FormSelectCurrency}
-        component={SwapFormSelectCurrency}
-        options={{
-          headerTitle: () => <StepHeader title={t("transfer.swap.form.to")} />,
-          headerRight: () => null,
-        }}
-      />
-      <Stack.Screen
-        name={ScreenName.SwapFormSelectProviderRate}
-        component={SwapFormSelectProviderRate}
-        options={{
-          headerTitle: () => (
-            <StepHeader title={t("transfer.swap.form.summary.method")} />
-          ),
-          headerRight: () => null,
-        }}
-      />
-      <Stack.Screen
-        name={ScreenName.SwapV2FormSelectFees}
-        component={SwapFormSelectFees}
-        options={{
-          headerTitle: () => (
-            <StepHeader title={t("transfer.swap.form.summary.fees")} />
-          ),
-          headerRight: () => null,
-        }}
+        options={{ headerShown: false }}
       />
       <Stack.Screen
         name={NavigatorName.Lending}
@@ -565,7 +511,7 @@ export default function BaseNavigator() {
       />
       <Stack.Screen
         name={ScreenName.Account}
-        component={Account}
+        component={readOnlyModeEnabled ? ReadOnlyAccount : Account}
         options={{ headerShown: false }}
       />
       <Stack.Screen
@@ -650,6 +596,7 @@ export default function BaseNavigator() {
         component={BleDevicePairingFlow}
         options={{
           title: "",
+          headerShown: false,
         }}
       />
       <Stack.Screen