+++ apps/ledger-live-mobile/src/locales/index.ts
@@ -28,3 +28,6 @@ exports.zh = {
 exports.pt = {
   common: require("./pt-BR/common.json"),
 };
+exports.ar = {
+  common: require("./ar/common.json"),
+};