+++ libs/ledger-live-common/src/apps/mock.ts
@@ -11,7 +11,7 @@ import {
 import { findCryptoCurrency } from "../currencies";
 import type { ListAppsResult, AppOp, Exec, InstalledItem } from "./types";
 import { getBTCValues } from "../countervalues/mock";
-import { DeviceModelId } from "@ledgerhq/devices";
+import { DeviceModelId, identifyTargetId } from "@ledgerhq/devices";
 import { App, DeviceInfo, FinalFirmware } from "@ledgerhq/types-live";
 
 export const deviceInfo155 = {
@@ -25,6 +25,40 @@ export const deviceInfo155 = {
   majMin: "1.5",
   targetId: 823132164,
 };
+
+export const deviceInfo210lo5: DeviceInfo = {
+  bootloaderVersion: "1.16",
+  hardwareVersion: 0,
+  isBootloader: false,
+  isOSU: false,
+  isRecoveryMode: false,
+  languageId: 0,
+  majMin: "2.1",
+  managerAllowed: false,
+  mcuVersion: "2.30",
+  onboarded: true,
+  pinValidated: true,
+  providerName: null,
+  seTargetId: 855638020,
+  seVersion: "2.1.0-lo5",
+  targetId: 855638020,
+  version: "2.1.0-lo5",
+};
+
+export const deviceInfo202 = {
+  version: "2.0.2",
+  isBootloader: false,
+  isOSU: false,
+  managerAllowed: true,
+  mcuVersion: "2.30",
+  pinValidated: true,
+  providerName: null,
+  majMin: "2.0",
+  targetId: 855638020,
+  seVersion: "2.0.2",
+  seTargetId: 855638020,
+};
+
 const firmware155: FinalFirmware = {
   id: 24,
   name: "1.5.5",
@@ -145,7 +179,9 @@ export function mockListAppsResult(
     appByName,
     appsListNames: apps.map((a) => a.name),
     deviceInfo,
-    deviceModelId: <DeviceModelId>"nanoS",
+    deviceModelId: deviceInfo.seTargetId
+      ? identifyTargetId(deviceInfo.seTargetId)?.id ?? <DeviceModelId>"nanoS"
+      : <DeviceModelId>"nanoS",
     firmware: firmware155,
     installed,
     installedAvailable: true,